requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic-settings>=2.2
SQLAlchemy>=2.0
alembic>=1.13
psycopg[binary]>=3.1
celery[redis]>=5.3
redis>=5.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
bcrypt<5
python-multipart>=0.0.9
structlog>=24.0
